+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Crash when array destructuring is exist ([#77](https://github.com/esdoc/esdoc/pull/77) Thanks @noraesae, [#76](https://github.com/esdoc/esdoc/issues/76))
   - Crash when computed property method is exist ([#78](https://github.com/esdoc/esdoc/pull/78) Thanks @noraesae, [#73](https://github.com/esdoc/esdoc/issues/73))
   - Crash when loading config without .js and .json ([#74](https://github.com/esdoc/esdoc/issues/74))
+  - Crash when unknown class new expression variable is exist ([#75](https://github.com/esdoc/esdoc/issues/75))
 
 ## 0.2.5 (2015-09-06)
 - **Feat**